migrate to configurable/flexible DefaultTwitterAuthDao

+++ src/groovy/com/the6hours/grails/springsecurity/twitter/TwitterAuthToken.groovy
@@ -12,7 +12,7 @@ import org.springframework.security.authentication.AbstractAuthenticationToken
  */
 class TwitterAuthToken extends AbstractAuthenticationToken implements Authentication {
 
-    int userId
+    long userId
     String screenName
     String tokenSecret
     String token
